实现PowerShell脚本权限认证与安全检测；给EffectBundle实现SettableFields，允许编编程化配置；优化几个和项目交互的cmdlet实现；修复XAMLCodeblockRenderer里的严重安全漏洞和Android光栅化的小bug

+++ projectFrameCut.Render.HwAccelEngine/Platforms/Android/OpenGLVectorRasterizer.cs
@@ -373,9 +373,9 @@ namespace projectFrameCut.Render.HwAccelEngine.Platforms.Android
                     {{'#'}}endif
 
                     outPacked[i] = pr;
-                    outPacked[i + {{pixelCount}}] = pg;
-                    outPacked[i + {{pixelCount * 2}}] = pb;
-                    outPacked[i + {{pixelCount * 3}}] = pa;
+                    outPacked[i + uint({{pixelCount}})] = pg;
+                    outPacked[i + uint({{pixelCount * 2}})] = pb;
+                    outPacked[i + uint({{pixelCount * 3}})] = pa;
                 }
                 """;
         }
