+++ projectFrameCut.Render.HwAccelEngine/Platforms/Android/VulkanVectorRasterizer.cs
@@ -318,9 +318,9 @@ void main()
     {{'#'}}endif
 
     outPacked[i] = pr;
-    outPacked[i + {{pixelCount}}] = pg;
-    outPacked[i + {{pixelCount * 2}}] = pb;
-    outPacked[i + {{pixelCount * 3}}] = pa;
+    outPacked[i + uint({{pixelCount}})] = pg;
+    outPacked[i + uint({{pixelCount * 2}})] = pb;
+    outPacked[i + uint({{pixelCount * 3}})] = pa;
 }
 """;
         }
